requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
PyYAML==6.0
requests==2.28.2
sentry-sdk==1.16.0
urllib3==1.26.14
